cs_api/requirements.txt
pyyaml
flask
flask_cors
uwsgi
schedule